requirements.txt
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pycrypto==2.6.1
pycryptodome==3.12.0
pyflakes==2.4.0
scapy==2.4.5
